Train more nets
