+++ src/agent/ids.json
@@ -1,0 +1,12 @@
+{
+    "Diluc":0,
+    "LargeWindSpirit":1,
+    "Sucrose":2,
+    "GlacialWaltz":3,
+    "Kaeya":4,
+    "Oz":5,
+    "Fischl":6,
+    "Senho":7,
+    "Frostflake":8,
+    "Ayaka":9
+}
